move conjFunctor to complex_functor.h
