+++ paddle/fluid/operators/conj_op.h
@@ -17,48 +17,12 @@
 #include "paddle/fluid/framework/eigen.h"
 #include "paddle/fluid/framework/op_registry.h"
 #include "paddle/fluid/framework/operator.h"
+#include "paddle/fluid/operators/math/complex_functors.h"
 #include "paddle/fluid/platform/for_range.h"
 
 namespace paddle {
 namespace operators {
 using Tensor = framework::Tensor;
-
-template <typename T>
-using EnableComplex =
-    typename std::enable_if<std::is_same<T, platform::complex64>::value ||
-                            std::is_same<T, platform::complex128>::value>::type;
-
-template <typename T>
-using DisableComplex = typename std::enable_if<
-    !std::is_same<T, platform::complex64>::value &&
-    !std::is_same<T, platform::complex128>::value>::type;
-
-template <typename T, typename Enable = void>
-struct ConjFunctor;
-
-template <typename T>
-struct ConjFunctor<T, EnableComplex<T>> {
-  ConjFunctor(const T* input, int64_t numel, T* output)
-      : input_(input), numel_(numel), output_(output) {}
-
-  HOSTDEVICE void operator()(size_t idx) const {
-    output_[idx] = T(input_[idx].real, -input_[idx].imag);
-  }
-  const T* input_;
-  int64_t numel_;
-  T* output_;
-};
-
-template <typename T>
-struct ConjFunctor<T, DisableComplex<T>> {
-  ConjFunctor(const T* input, int64_t numel, T* output)
-      : input_(input), numel_(numel), output_(output) {}
-
-  HOSTDEVICE void operator()(size_t idx) const { output_[idx] = input_[idx]; }
-  const T* input_;
-  int64_t numel_;
-  T* output_;
-};
 
 template <typename DeviceContext, typename T>
 class ConjKernel : public framework::OpKernel<T> {
@@ -74,7 +38,7 @@ class ConjKernel : public framework::OpKernel<T> {
 
     auto& dev_ctx = context.template device_context<DeviceContext>();
     platform::ForRange<DeviceContext> for_range(dev_ctx, numel);
-    ConjFunctor<T> functor(x_data, numel, out_data);
+    math::ConjFunctor<T> functor(x_data, numel, out_data);
     for_range(functor);
   }
 };
