+++ python/paddle/fluid/tests/unittests/white_list/check_shape_white_list.py
@@ -21,7 +21,6 @@ NEED_TO_FIX_OP_LIST = [
     'lstmp',
     'margin_rank_loss',
     'matmul',
-    'matmul_v2',
     'scatter',
     'soft_relu',
     'squared_l2_distance',
